backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy[asyncio]
asyncpg
httpx
python-dotenv
python-dateutil